hr-agent/requirements.txt
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.3.0
python-dotenv>=1.0.0
